Run builtins from a struct array
The new Builtin struct contains the name of the builtin as a string and
a function pointer which should run when the builtin is called. An array
of this struct is used to store all the available builtins with each
one's name and respective function. This array is then looped through
comparing the name value to the first token, calling the function stored
in the struct when they equal each other.